+++ tdt/flash/common/fup.src/crc32.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-uint64_t crc32(uint64_t crc, uint8_t *buf, uint32_t size);
+uint32_t crc32(uint32_t crc, uint8_t *buf, uint32_t size);
 
 #endif /* CRC32_H_ */
